Add support for loading files inside str2s, inside vpps
This is a first draft of this code. It's complete garbage code, but it works so far. This was hard for me to conceptualize so I started off just making it work. I'll rewrite it to be cleaner and more concise, ideally using FileRefs for everything instead of having multiple lists and wrangling a bunch of strings. I'll likely also move most of the streaming caching stuff into a new class, like CacheManager or StreamingManager.