Fix several caching and threading bugs
Wrapped HandleOpenFileEvent in MainView.xaml.cs in  `Application.Current.Dispatcher.Invoke()` since it fixes an issue with threading conflicts. Also improved some of the the code in ProjectManager slightly so that `IsFileCached` fails in fewer cases, and so TryGetFile works better. ProjectManager is still an ugly chunk of code, so I'll rewrite it at some point to be more legible, and less buggy.