Replace FileExplorer for loops with foreach loops
I tested the initial file explorer list generation time for manual for loops and foreach loops, with 20 samples each. The manual for loops on average took 421.2ms, and the foreach loops on average took 401.8ms. The difference in generation time in negligible, and foreach loops make the code much cleaner and easy to read, so I'm switching this code to use foreach loops.