Rewrite CacheManager API to be less insane
Previously made use of FileExplorerItemViewModel on a few API functions. This made it so anything that wanted to use it needed awareness of the file explorer tree, which causes a lot of problems. Now it uses filename strings, and CacheFiles for API functions instead, which makes it easier for anything to use the API. 

I still think the CacheManager code is unwieldy and this made it ~100 LOC longer, but it's an improvement.