Add full file explorer alphabetical sorting, update submodule with stream handling improvements
Forgot to update the submodule 2 weeks ago with some important improvements in stream handling so they are properly disposed. Also made a change to Packfiles that makes it easier to sort their subfiles. Used that change to alphabetically sort depth 1 files. 

Doesn't sort depth > 1 files yet. There are usually < 10 of them in a group so it's not really worth the extra code.