Temporary fix for incorrect asm_pc data
Turns out that asm_pc file for interfaces.vpp_pc//badges_bitmap_sheet.str2_pc having incorrect data about it's contents. The asm_pc file claims that it contains a file called "interface-badges.cvbm_pc" when it really contains "interface-badges.cpeg_pc". 

These are equivalent when it comes to loading the texture data, and both valid texture names. The problem is that CacheManager expects names to match exactly and considers files with different extensions to not be equal, causing loading of this file to fail. 

This patch just adds a special case to fix this files name. If it turns out that there are many other files with this issue some other solution will need to be made. This works for now though.